pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffviews"
version = "0.1.0"
description = "Diffusion model activation visualizer with model-agnostic adapter interface"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "DMD2 Team"}
]
keywords = ["diffusion", "visualization", "umap", "activations", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "pillow>=9.0.0",
    "scikit-learn>=1.0.0",
    "umap-learn>=0.5.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
viz = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "plotly>=5.18.0",
]
lmdb = [
    "lmdb>=1.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "diffviews[viz,lmdb,dev]",
]

[project.urls]
Repository = "https://github.com/your-org/diffviews"

[project.scripts]
diffviews = "diffviews.scripts.cli:main"

[project.entry-points."diffviews.adapters"]
# Adapters register themselves here
# Example: imagenet-64 = "diffviews_dmd2:DMD2ImageNetAdapter"

[tool.setuptools.packages.find]
where = ["."]
include = ["diffviews*"]

[tool.setuptools.package-data]
diffviews = ["py.typed"]
"*" = ["*.json"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
